Make serde classes public
